Code cleanup Project: - Reduced ::data namespace nesting - Marked applicable variables const - Resolved some type conversion warnings - Replaced some numbers with constants - Removed deprecated functions with 0 usages - Merged headers defining a single(few) function(s)
Tests:
- Moved CMake test source files into local CMakeLists.txt
- Removed underscores from test suites
- Grouped common data into test fixtures
- Changed J_TEST_USE_ONLY to #ifdef #endif blocks
- Attempted to replace macros with methods
- Commented out long running tests

+++ include/data/prototype/entity/entity.h
@@ -53,8 +53,8 @@ namespace jactorio::data
 
 
 		// Number of tiles this entity spans
-		PYTHON_PROP_REF_I(Entity, unsigned short, tile_width, 1)
-		PYTHON_PROP_REF_I(Entity, unsigned short, tile_height, 1)
+		PYTHON_PROP_REF_I(Entity, uint8_t, tile_width, 1)
+		PYTHON_PROP_REF_I(Entity, uint8_t, tile_height, 1)
 
 		// Can be rotated by player?
 		PYTHON_PROP_REF_I(Entity, bool, rotatable, false)
